+++ Chaos.NaCl/Chaos.NaCl.csproj
@@ -42,9 +42,12 @@
     <Compile Include="Internal\Array8.cs" />
     <Compile Include="Internal\ByteIntegerConverter.cs" />
     <Compile Include="CryptoBytes.cs" />
+    <Compile Include="Internal\Ed25519Ref10\elligator.cs" />
     <Compile Include="Internal\Ed25519Ref10\fe_bytesLE.cs" />
     <Compile Include="Internal\Ed25519Ref10\fe_cswap.cs" />
     <Compile Include="Internal\Ed25519Ref10\fe_mul121666.cs" />
+    <Compile Include="Internal\Ed25519Ref10\chi.cs" />
+    <Compile Include="Internal\Ed25519Ref10\q58.cs" />
     <Compile Include="Internal\Ed25519Ref10\scalarmult.cs" />
     <Compile Include="Internal\Ed25519Ref10\sc_clamp.cs" />
     <Compile Include="Internal\InternalAssert.cs" />
